split header files
