+++ include/functions.h
@@ -1,16 +1,11 @@
 #ifndef _FUNCTIONS_H
 #define _FUNCTIONS_H
 
-#include "ultra64.h"
-#include "common_structs.h"
-#include "variables.h"
-
 void bootproc(void);
 void idle(void * arg);
 void mainproc(void * arg);
 
 void test_write(void);
-void init_memory(void);
 void main_no_expPak(void);
 void main_expPak(void);
 void threadTimerTick(void);
@@ -25,29 +20,7 @@ void replace_loaded_dll_list(DLLInst arg0[], s32 arg1);
 
 s32 read_file_region(u32 id, void *dst, u32 offset, s32 size);
 
-void free(void* p);
 DLLFile * dll_load_from_tab(u16, u32 *);
 void dll_relocate(DLLFile* dll);
-
-void *malloc(s32 size, s32 tag, const char *name);
-void _bcopy(const void *src,void *dst,int length);
-
-void matrix_from_srt(MtxF *mf, SRT *srt);
-
-// theta: [-32768..32768) => [-pi..pi)
-// returns: [-65536..65536] => [-1..1]
-s32 cos16_precise(s16 theta);
-
-// theta: [-32768..32768) => [-pi..pi)
-// returns: [-65536..65536] => [-1..1]
-s32 sin16_precise(s16 theta);
-
-// theta: [-32768..32768) => [-pi..pi)
-// returns: [-65536..65536] => [-1..1]
-s32 cos16(s16 theta);
-
-// theta: [-32768..32768) => [-pi..pi)
-// returns: [-65536..65536] => [-1..1]
-s32 sin16(s16 theta);
 
 #endif
