+++ include/sys/audio/audio.h
@@ -1,0 +1,4 @@
+/** Audio system
+ */
+
+extern u8 alSynFlag;
