+++ include/sys/controller.h
@@ -1,0 +1,10 @@
+/** Controller/gamepad system
+ */
+
+/**
+ * The index of the last inserted controller.
+ * 
+ * For example, if one controller is inserted then this will be 0.
+ * If no controllers are inserted, this will be -1.
+ */
+extern s8 gLastInsertedControllerIndex;
