+++ include/sys/crash.h
@@ -1,0 +1,9 @@
+/** Crash handler system
+ */
+
+//use for crash screens
+struct ErrString{
+	u32 code1;
+	u32 code2;
+	char* text;
+};
