+++ include/sys/debug.h
@@ -1,0 +1,4 @@
+/** Misc debug stuff
+ */
+
+#define DEBUG_LOOP1 do {/* omitted debug code */} while(0)
